Expose seekTo method on event video
