+++ src/components/Details/EventVideo/EventVideo.stories.tsx
@@ -1,7 +1,7 @@
-import React from "react";
+import React, { createRef } from "react";
 import { Story, Meta } from "@storybook/react";
 
-import EventVideo, { EventVideoProps } from "./EventVideo";
+import EventVideo, { EventVideoProps, EventVideoRef } from "./EventVideo";
 
 export default {
   component: EventVideo,
@@ -13,4 +13,5 @@ const Template: Story<EventVideoProps> = (args) => <EventVideo {...args} />;
 export const Default = Template.bind({});
 Default.args = {
   uri: "https://video.seattle.gov/media/council/council_113020_2022091V.mp4",
+  componentRef: createRef<EventVideoRef>(),
 };
